Add option to increase predictability of thumbnail filenames
Thumbnails are named by their file of origin with page number as
suffix. By default files are renamed such that the page suffix contains
the same number of characters for all thumbnails (01,02,…,10,11). As a
result the first thumbnail could either be called <somename>1.ext or
<somename>01.ext or even <somename>001.ext, making it harder to pick up
these files downstream. Here an option is added to change this
behaviour so that the first thumbnail is always called <somename>1.ext